requirements.txt
Flask==2.1.2
requests==2.28.1
urllib3==1.26.10
lxml==4.9.1
pandas==1.4.3
gunicorn~=20.1.0
PyGithub~=1.55
google-cloud-secret-manager==2.11.1
google-cloud-logging==3.1.2
sendgrid~=6.9.3
fake-useragent~=0.1.11
pytz==2022.1
python-dotenv==0.21.0